Patch jay_std_clock
I think this finally fixes it for good?

+++ jaylib/jaylib.h
@@ -1980,7 +1980,7 @@ static
 jay_value
 jay_std_clock(jay_value *args, jay_closure *closure) {
 	clock_t time = clock();
-	double seconds = (time / CLOCKS_PER_SEC);
+	double seconds = (time / (double)CLOCKS_PER_SEC);
 	return jay_box_number(seconds);
 }
